feat: Update README.md with new images
This commit updates the README.md file to fix the indentation of the
images and adds new images for Obsidian, and abandoned
languages and tools.

The changes were made to improve the visual presentation of the README
file and provide accurate representation of the technologies and tools
used in the project.